Convert location events from emitter to postal
I'm attempting to implement a somewhat pure event sourcing approach to the
game. Postal provides a message broker, so I can decouple the publishers
from the handlers. The idea is that the handlers do not change the state
of the game--they merely publish events. Both the text engine and the
state objects will subscribe to these events. The state objects will now
update themselves based on the data in the event, and the text engine must
rely solely on the data in the event message instead of the game state
objects, since I'm not sure I can guarantee which object will respond to
the events first.